start to refacor teacher into module
